Add fzf vim plugin file for macOS
To work properly on macOS, the vim fzf plugin requires "plugging" a fzf
file. This commit adds this file to vim/plugins.vim and "plug" it
depending on OS type.

Also, I'm removing vim ALE setup for Kotlin language server, since I did
not have a good experience with it.