Split ~/.profile into OS specific files
The ~/.profile included commands only available in Linux (setxkbmap) and
it is not loaded in the same way in macOS. In fact, ~/.profile was just
a symbolic link to ~/.bashrc in my macOS (I guess I did it when I first
installed my dotfiles on it).

Now, the ~/.profile is the "main" file, which loads OS specific files,
which are installed through the main installation script. The file at
`bash/profile.linux` is the old ~/.profile now and I am creating another
for macOS.

P.S.1: Somehow, editing files from terminal in vim "broke", it was not
using UTF8 encoding, but `latin`. Now I explicitly set the language
encoding through environment variables.

P.S.2: The logic for setting the system theme based on the current time
was a bit flawed, the time formatting for getting the current time
as a integer was wrong, it was getting hour and hour instead of hour and
minute. I fixed it.