Add startup script for system theme on macOS
As macOS does not load the ~/.profile in the same way Linux does (every
new shell is login shell), exporting the "$SYSTEM_THEME" environment
variable does not work properly: it is not available for GUI
applications, so starting kitty from the GUI will not load the desired
color theme.

It is a bit of arcane magic to make it happen on macOS: you need to
create a launch agent to be loaded on startup (after macOS GUI login),
which includes creating a '.plist' file (XML) and configure it to run a
script. To export environment variables globally in macOS, you need to
use the `launchctl setenv ENVVAR value` command. After, you load the
agent into launchd, enabling it to always run at startup.

All these steps are configured in 'install/macos.sh'.